requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==2.1.5
django-cors-headers==3.4.0
django-rq==2.3.2
djangorestframework==3.11.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
pycodestyle==2.6.0
pylint==2.5.3
pytz==2020.1
qr==0.6.0
redis==3.5.3
requests==2.24.0
rq==1.5.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
hdacpy
django-sslserver